Fix Native::Response#headers. Now returns the cloned Net::HTTP::Response itself.
The original implementation returned a @http_response.to_hash. However this made
the headers a case-sensitive Hash. This is against RFC 2626 and incompatibile with the Net::HTTP HTTPHeader object class too.

The code this commit replaced was my attempt to enforce uniformity for the hash keys. Was not ideal - it still broke the RFC.