added custom equality and hash to Native::Response.
This is required for the tests that deal with caching (get_spec). In these tests, a response is first received, and it is assumed it is cached automatically.  The test then again requests the same uri and expects the old response to be made available from the cache. Comparing the response that was obtained from the cache with the first response should yield a true. However, if the cache stored is a 'clone' of the original response, the default comparison would say they are different. This comparison was implemented for this use case.